requirements.txt
asgiref==3.2.3
Django==3.0.3
pytz==2019.3
sqlparse==0.3.1
djangorestframework>=3.10.0,<3.11.0
psycopg2-binary>=2.8.0
flake8>=3.6.0,<3.7.0